Added political parties controller

+++ DocumentDataAPI/DocumentDataAPI/Data/Repositories/IBiasWordCountRepository.cs
@@ -4,6 +4,6 @@ namespace DocumentDataAPI.Data.Repositories;
 
 public interface IBiasWordCountRepository : IRepository<BiasWordCountModel>
 {
-    Task<IEnumerable<int>> AddBatch(List<BiasWordCountModel> models);
-    Task<int> DeleteAll();
+    Task<IEnumerable<long>> AddBatch(List<BiasWordCountModel> models);
+    Task<long> DeleteAll();
 }
